requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
python-dotenv>=1.0
yt-dlp
pymysql
